Updating Silex provider to use file prefix.
If debugging is enabled, the "dev." prefix will be used.  Otherwise, the "prod." prefix is used.  Because of the way Wisdom handles file prefixes, use of it is not required.  However, by not using a prefix in your file name, you allow it to be used in both debugging and non-debugging modes.